+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Make sure stdout and stderr output is recorded.
 
+* Try to find locally-install phantomjs first.
+
 
 # 1.0.3