+++ NEWS.md
@@ -1,4 +1,4 @@
-# 1.0.4.9000
+# 1.0.5
 
 * Make sure stdout and stderr output is recorded.